+++ pypodo/list.py
@@ -8,6 +8,7 @@ from pypodo.properties import (
     REGEX_SPACE_OR_ENDLINE,
 )
 from pypodo.config import (
+    read_config_boolean,
     todofilefromconfig,
 )
 from pypodo.args import compute_args
@@ -28,12 +29,8 @@ def listtask(openfile=open):
     with openfile(todofilefromconfig(), "r") as todofile:
         for line in todofile.readlines():
             # without filter -> we print all
-            if not compute_args().filter:
-                data.append(printlinetodo(line))
-                empty = False
-            # with filter -> we check tag
-            else:
-                display = True
+            display = True
+            if compute_args().filter:
                 for tagtofilter in compute_args().filter:
                     # regex to search tags "#toto " or "#toto" at the end of the line
                     if not re.findall(
@@ -43,9 +40,19 @@ def listtask(openfile=open):
                         line.rstrip("\n"),
                     ):
                         display = False
-                if display:
-                    data.append(printlinetodo(line))
-                    empty = False
+            if compute_args().exclude:                       
+                for tagtoexclude in compute_args().exclude:
+                    # regex to search tags "#toto " or "#toto" at the end of the line
+                    if re.findall(
+                        " #"
+                        + re.escape(tagtoexclude)
+                        + REGEX_SPACE_OR_ENDLINE,
+                        line.rstrip("\n"),
+                    ):
+                        display = False                        
+            if display:
+                data.append(printlinetodo(line))
+                empty = False
     if empty:
         if compute_args().filter:
             printwarning("the filtered todolist is empty")
@@ -55,5 +62,8 @@ def listtask(openfile=open):
         if compute_args().condensate:
             table = columnar(data, no_borders=True, wrap_max=0)
         else:
-            table = columnar(data, headers, no_borders=False, wrap_max=0)            
+            if (read_config_boolean("FONCTIONAL", "condensate", "False")== "True"):
+                table = columnar(data, no_borders=True, wrap_max=0)
+            else:    
+                table = columnar(data, headers, no_borders=False, wrap_max=0)            
         print(table)