+++ pypodo/properties.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-
-STR_PATH_HOME__TODORC_ = str(Path.home()) + "/.todo.rc"
+DOSSIER_CONFIG_PYPODO="pypodo"
+TODO_RC_FILE = "config"
 REGEX_INDEX = "^\\d+$"
 REGEX_SPACE_OR_ENDLINE = "( |$)"